refactored utils into a common object
