+++ parser/src/main/scala/org/polystat/odin/parser/Utils.scala
@@ -1,9 +1,9 @@
-package org.polystat.odin.parser.fastparse
+package org.polystat.odin.parser
 
 import com.github.tarao.nonempty.collection.NonEmpty
-import org.polystat.odin.core.ast.{EOAnonExpr, EOBnd, EOBndExpr, EOCopy, EODot, EOExpr}
-import org.polystat.odin.core.ast.astparams.EOExprOnly
 import higherkindness.droste.data.Fix
+import org.polystat.odin.core.ast.astparams.EOExprOnly
+import org.polystat.odin.core.ast._
 
 object Utils {
   def createNonEmpty(objs: Seq[EOBnd[EOExprOnly]])
@@ -14,7 +14,7 @@ object Utils {
     }
   }
 
-   private def extractEOExpr(bnd: EOBnd[EOExprOnly]): EOExprOnly = {
+  private def extractEOExpr(bnd: EOBnd[EOExprOnly]): EOExprOnly = {
     bnd match {
       case EOAnonExpr(expr) => expr
       case EOBndExpr(_, expr) => expr
@@ -24,7 +24,7 @@ object Utils {
   // TODO: rewrite so that the information
   //  about names of bindings is not lost
   def createInverseDot(id: String,
-                               args: Vector[EOBnd[EOExprOnly]]): EOExprOnly = {
+                       args: Vector[EOBnd[EOExprOnly]]): EOExprOnly = {
     if (args.tail.nonEmpty) {
       Fix[EOExpr](
         EOCopy(
@@ -39,4 +39,3 @@ object Utils {
 
 
 }
-
