+++ parser/src/main/scala/org/polystat/odin/parser/scala_parser_combinators/Parser.scala
@@ -4,11 +4,11 @@ import com.github.tarao.nonempty.collection.NonEmpty
 import org.polystat.odin.core.ast._
 import org.polystat.odin.core.ast.astparams.EOExprOnly
 import errors._
-
 import higherkindness.droste.data.Fix
+import org.polystat.odin.parser.Utils.{createInverseDot, createNonEmpty}
 
 import scala.util.parsing.combinator.Parsers
-import scala.util.parsing.input.{ NoPosition, Position, Reader }
+import scala.util.parsing.input.{NoPosition, Position, Reader}
 
 object Parser extends Parsers {
   class WorkflowTokenReader(val tokens: Seq[Token]) extends Reader[Token] {
@@ -63,38 +63,6 @@ object Parser extends Parsers {
 
   private def meta: Parser[META] = {
     accept("meta", { case meta: META => meta })
-  }
-
-  private def createNonEmpty(
-    objs: Seq[EOBnd[EOExprOnly]]
-  ): NonEmpty[EOBnd[EOExprOnly], Vector[EOBnd[EOExprOnly]]] = {
-    NonEmpty.from(objs) match {
-      case Some(value) => value.toVector
-      case None => throw new Exception("1 or more arguments expected, got 0.")
-    }
-  }
-
-  private def extractEOExpr(bnd: EOBnd[EOExprOnly]): EOExprOnly = {
-    bnd match {
-      case EOAnonExpr(expr) => expr
-      case EOBndExpr(_, expr) => expr
-    }
-  }
-
-  private def createInverseDot(
-    id: IDENTIFIER,
-    args: Vector[EOBnd[EOExprOnly]]
-  ): EOExprOnly = {
-    if (args.tail.nonEmpty) {
-      Fix[EOExpr](
-        EOCopy(
-          Fix[EOExpr](EODot(extractEOExpr(args.head), id.name)),
-          createNonEmpty(args.tail)
-        )
-      )
-    } else {
-      Fix[EOExpr](EODot(extractEOExpr(args.head), id.name))
-    }
   }
 
   def commentsOrNewlines: Parser[List[Token]] =
@@ -198,7 +166,7 @@ object Parser extends Parsers {
     }
     val inverseDot = identifier ~ DOT ~ name ~ verticalApplicationArgs ^^ {
       case id ~ _ ~ name ~ args =>
-        EOBndExpr(name, createInverseDot(id, args))
+        EOBndExpr(name, createInverseDot(id.name, args))
     }
     val withArgs = singleLineApplication ~ name ~ verticalApplicationArgs ^^ {
       case target ~ name ~ args =>
@@ -214,7 +182,7 @@ object Parser extends Parsers {
     }
     val inverseDot = identifier ~ DOT ~ verticalApplicationArgs ^^ {
       case id ~ _ ~ args =>
-        EOAnonExpr(createInverseDot(id, args))
+        EOAnonExpr(createInverseDot(id.name, args))
     }
     val withArgs = singleLineApplication ~ verticalApplicationArgs ^^ {
       case target ~ args =>
